% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_stats.R
\name{get_player_stats}
\alias{get_player_stats}
\title{Gets player stats from the WoSo Stats database}
\usage{
get_player_stats(
  events,
  event_type,
  defending,
  poss_notes,
  def_notes,
  matches,
  lineups
)
}
\arguments{
\item{events}{A data frame representing the "events" table from the WoSo Stats
database.}

\item{event_type}{A data frame representing the "event_type" table from the WoSo
Stats database.}

\item{defending}{A data frame representing the "defending" table from the WoSo
Stats database.}

\item{poss_notes}{A data frame representing the "poss_notes" table from the WoSo
Stats database.}

\item{def_notes}{A data frame representing the "def_notes" table from the WoSo
Stats database.}

\item{matches}{A data frame representing the "matches" table from the WoSo Stats
database.}

\item{lineups}{A data frame representing the "lineups" table from the WoSo Stats
database.}
}
\value{
A data frame
}
\description{
\code{get_player_stats} creates a data frame representing a summary of advanced
stats for each player for each match in the WoSo Stats database. A glossary for
each metric is at \url{https://www.wosostats.com/glossary/}.
}
