% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Get a table from the WoSo Stats database}
\usage{
get_table(object)
}
\arguments{
\item{object}{The name of the object in the WoSo Stats database.}
}
\value{
A data frame representing \code{object}.
}
\description{
\code{get_table} creates a data frame for a specified table in the WoSo Stats
database. There are specific tables you can read, outlined below. Each table
is an Amazon S3 object which is ultimately, within \code{get_table}, read
as a csv file.
}
\examples{
dat <- get_table("events")
dat <- get_table("defending")
}
