% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_all.R
\name{get_table_all}
\alias{get_table_all}
\title{Get all tables from the WoSo Stats database}
\usage{
get_table_all()
}
\value{
A list of data frames.
}
\description{
\code{get_table_all} creates a list of data frames for all tables in the WoSo
Stats database. Each table is an Amazon S3 object which is ultimately,
within \code{get_table_all}, read as a csv file and assigned to a data frame.
}
\examples{
all_tables <- get_table_all()
}
