% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_match_log.R
\name{read_match_log}
\alias{read_match_log}
\title{Read an Excel match log}
\usage{
read_match_log(filename, matches, lineups)
}
\arguments{
\item{filename}{The filename of the Excel file in your working directory
representing the match log you want to read.}

\item{matches}{A data frame representing the "matches" table from the WoSo
Stats database.}

\item{lineups}{A data frame representing the "lineups" table from the WoSo
Stats database.}
}
\value{
A list of data frames
}
\description{
\code{read_match_log} reads a match log (in Excel format) in your working
directory and creates a list of data frames representing different types of
events in the match log. Match logs can be downloaded from a \href{https://drive.google.com/drive/folders/13-8Ws14GougTk_FZBv4k-VUCaRyml1hj?usp=sharing}{a Google Drive folder}
which is also linked in the README.
}
\examples{
\dontrun{
matches_tbl <- get_table("matches")
lineups_tbl <- get_table("lineups")
match_tables <- read_match_log(filename = "nwsl-2016-hou-orl-052016.xlsx",
matches = matches_tbl, lineups = lineups_tbl)
}
}
