% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multi_match_log.R
\name{read_multi_match_log}
\alias{read_multi_match_log}
\title{Reads multiple Excel match logs}
\usage{
read_multi_match_log(matches, lineups)
}
\arguments{
\item{matches}{A data frame representing the "matches" table from the WoSo
Stats database.}

\item{lineups}{A data frame representing the "lineups" table from the WoSo
Stats database.}
}
\value{
A list of data frames each representing various types of data sourced
from the Excel files in your working directory.
}
\description{
\code{read_multi_match_log} reads all the match logs (in Excel format) in your
working directory and creates a list of data frames aggregating the different
types of events in each match.
}
\examples{
\dontrun{
matches_tbl <- get_table("matches")
lineups_tbl <- get_table("lineups")
match_tables <- read_multi_match_log(matches = matches_tbl, lineups = lineups_tbl)
}
}
